+++ src/LogManager.cpp
@@ -21,7 +21,7 @@ std::string ErrorCode::getMessage()
 
 void LogManager::freeLogSpace()
 {
-    for (int i = entries.size() / 4; i >= 0; i--)
+    for (size_t i = entries.size() / 4; i >= 0; i--)
     {
         if (entries[i].number / 100 != 2)
         {
@@ -31,7 +31,7 @@ void LogManager::freeLogSpace()
     }
 }
 
-void LogManager::add(char letter, int number)
+void LogManager::add(char letter, uint16_t number)
 {
     entries.push_back(ErrorCode{letter, number});
     hasNewMessagesFlag = true;
@@ -45,7 +45,7 @@ void LogManager::add(char letter, int number)
 std::vector<std::string> LogManager::getNewMessages()
 {
     std::vector<std::string> newMessages;
-    for (int i = entries.size() - newMessageCount; i < entries.size(); i++)
+    for (size_t i = entries.size() - newMessageCount; i < entries.size(); i++)
     {
         newMessages.push_back(entries[i].getMessage());
     }
